create file index in FileTable
